// src/webparts/pixExecutivePortal/models/IEventSettings.ts
export interface IEventSettings {

  Id: number;

  Title: string;

  EventName: string;

  EventSubTitle: string;

  EventLocation: string;

  EventDate: string;

  HeroTitleLine1: string;

  HeroTitleLine2: string;

  InvitationLabel: string;

  EventDuration: string;

  Description: string;

  IsActive: boolean;
}